docker/consumer/requirements.txt
apache_beam
beam_nuggets
kafka-python
scikit-learn
nltk
happybase
joblib